Add body hash to LTI outcome message.
This is necessary to properly implement the LTI outcome passback
spec. It was not included previously because it was causing problems
with Canvas, but Blackboard will not accept outcomes unless they are
properly signed.

Note that this commit sends the response with verify=False because
the test sandbox we're working against has a self-signed cert. This
should be removed before we actually merge this into master.