Remove the edx.org comprehensive theme
There's a lot of useful things in PR #8271 that provide a framework
for the comprehensive theming system. If we need to remove the edx.org
theme from the codebase, we can leave most of PR #8271 in, to make it
easier to build on top of and get this feature back in.