Removed unnecessary editable flag for requirements
Installing requirements with the editable flag implies that the libraries will be further developed locally, thus they are placed in a special directory. When we eventually move away from Git-based requirements, we are forced to rebuild AMIs from scratch or run special scripts to uninstall these editable packages locally. This is an unnecessary burden given how infrequently we actually need editable packages.