Fix logic flaw in checking feature flag
Add unit tests to check the 4 combined states of our 2 feature flags.
One can configure the LMS to return no history records, but they will
continue to be saved.

This also hopefully improves some of the comments around the feature
flags in common.py and the offset for the tables.